Minor changes, mostly float stuff. VS2019 solution and project upgrades.

+++ coop/g_combat.c
@@ -790,11 +790,11 @@ CheckArmor(edict_t *ent, vec3_t point, vec3_t normal, int damage,
 
 	if (dflags & DAMAGE_ENERGY)
 	{
-		save = ceil(((gitem_armor_t *)armor->info)->energy_protection * damage);
+		save = ceilf(((gitem_armor_t *)armor->info)->energy_protection * damage);
 	}
 	else
 	{
-		save = ceil(((gitem_armor_t *)armor->info)->normal_protection * damage);
+		save = ceilf(((gitem_armor_t *)armor->info)->normal_protection * damage);
 	}
 
 	if (save >= client->pers.inventory[index])
