+++ coop/g_func.c
@@ -508,7 +508,7 @@ plat_CalcAcceleratedMove(moveinfo_t *moveinfo)
 		float f;
 
 		f = (moveinfo->accel + moveinfo->decel) / (moveinfo->accel * moveinfo->decel);
-		moveinfo->move_speed = (-2 + sqrt(4 - 4 * f * (-2 * moveinfo->remaining_distance))) / (2 * f);
+		moveinfo->move_speed = (-2 + sqrtf(4 - 4 * f * (-2 * moveinfo->remaining_distance))) / (2 * f);
 		decel_dist = AccelerationDistance(moveinfo->move_speed, moveinfo->decel);
 	}
 
@@ -596,7 +596,7 @@ plat_Accelerate(moveinfo_t *moveinfo)
 		   and cross over the decel_distance; figure the average speed for the
 		   entire move */
 		p1_distance = moveinfo->remaining_distance - moveinfo->decel_distance;
-		p1_speed = (old_speed + moveinfo->move_speed) / 2.0;
+		p1_speed = (old_speed + moveinfo->move_speed) / 2.0f;
 		p2_distance = moveinfo->move_speed * (1.0 - (p1_distance / p1_speed));
 		distance = p1_distance + p2_distance;
 		moveinfo->current_speed = (p1_speed * (p1_distance /
@@ -4465,14 +4465,14 @@ SP_func_door_secret(edict_t *ent)
 
 	if (ent->spawnflags & SECRET_1ST_DOWN)
 	{
-		width = fabs(DotProduct(up, ent->size));
+		width = fabsf(DotProduct(up, ent->size));
 	}
 	else
 	{
-		width = fabs(DotProduct(right, ent->size));
+		width = fabsf(DotProduct(right, ent->size));
 	}
 
-	length = fabs(DotProduct(forward, ent->size));
+	length = fabsf(DotProduct(forward, ent->size));
 
 	if (ent->spawnflags & SECRET_1ST_DOWN)
 	{
