+++ coop/g_monster.c
@@ -645,7 +645,7 @@ M_WorldEffects(edict_t *ent)
 				/* drown! */
 				if (ent->pain_debounce_time < level.time)
 				{
-					dmg = 2 + 2 * floor(level.time - ent->air_finished);
+					dmg = 2 + 2 * floorf(level.time - ent->air_finished);
 
 					if (dmg > 15)
 					{
@@ -669,7 +669,7 @@ M_WorldEffects(edict_t *ent)
 				/* suffocate! */
 				if (ent->pain_debounce_time < level.time)
 				{
-					dmg = 2 + 2 * floor(level.time - ent->air_finished);
+					dmg = 2 + 2 * floorf(level.time - ent->air_finished);
 
 					if (dmg > 15)
 					{
